Perbaiki proses hapus opd, beri validasi pada saat proses hapus opd dan program cek ada tidak data child relasi nya

+++ src/main/java/cc/kertaskerja/bontang/opd/domain/OpdService.java
@@ -1,15 +1,22 @@
 package cc.kertaskerja.bontang.opd.domain;
 
+import cc.kertaskerja.bontang.kegiatan.domain.KegiatanRepository;
 import cc.kertaskerja.bontang.opd.domain.exception.OpdAlreadyExistException;
+import cc.kertaskerja.bontang.opd.domain.exception.OpdDeleteForbiddenException;
 import cc.kertaskerja.bontang.opd.domain.exception.OpdNotFoundException;
+import cc.kertaskerja.bontang.program.domain.ProgramRepository;
 import org.springframework.stereotype.Service;
 
 @Service
 public class OpdService {
     private OpdRepository opdRepository;
+    private ProgramRepository programRepository;
+    private KegiatanRepository kegiatanRepository;
 
-    public OpdService(OpdRepository opdRepository) {
+    public OpdService(OpdRepository opdRepository, ProgramRepository programRepository, KegiatanRepository kegiatanRepository) {
         this.opdRepository = opdRepository;
+        this.programRepository = programRepository;
+        this.kegiatanRepository = kegiatanRepository;
     }
 
     public Iterable<Opd> findAll() {
@@ -41,6 +48,16 @@ public class OpdService {
     public void hapusOpd(String kodeOpd) {
         if (!opdRepository.existsByKodeOpd(kodeOpd)) {
             throw new OpdNotFoundException(kodeOpd);
+        }
+
+        boolean opdDigunakanPadaProgram = programRepository.existsByKodeOpd(kodeOpd);
+        if (opdDigunakanPadaProgram) {
+            throw new OpdDeleteForbiddenException(kodeOpd);
+        }
+
+        boolean opdDigunakanPadaKegiatan = kegiatanRepository.existsByKodeOpd(kodeOpd);
+        if (opdDigunakanPadaKegiatan) {
+            throw new OpdDeleteForbiddenException(kodeOpd);
         }
 
         opdRepository.deleteByKodeOpd(kodeOpd);
